hellodjango/requirements.txt
asgiref==3.9.1
channels==4.3.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
langchain_community==0.4.1
langchain_core==1.0.4
langchain_huggingface==1.0.1
openai==2.7.2
